fix(docs/examples): move utils out of lib
