+++ docs/examples/rust/stardust/alias-migration.rs
@@ -9,7 +9,7 @@
 use std::str::FromStr;
 
 use anyhow::{anyhow, Result};
-use docs_examples::{clean_keystore, publish_custom_nft_package, setup_keystore};
+use docs_examples::utils::{clean_keystore, publish_custom_nft_package, setup_keystore};
 use iota_keys::keystore::AccountKeystore;
 use iota_sdk::{
     rpc_types::{IotaData, IotaObjectDataOptions, IotaTransactionBlockResponseOptions},
