+++ docs/examples/rust/stardust/foundry-output-claim.rs
@@ -6,7 +6,7 @@
 //! generated from iota-genesis-builder/src/stardust/test_outputs.
 
 use anyhow::anyhow;
-use docs_examples::{clean_keystore, fund_address, setup_keystore};
+use docs_examples::utils::{clean_keystore, fund_address, setup_keystore};
 use iota_keys::keystore::AccountKeystore;
 use iota_sdk::{
     rpc_types::{
